Update README.md and publish LICENSE.txt as well as project resources
For transparancy, the project license as well as  all of the text file resources have been pubished.
This will allow end users to better understand the customization available to them prior to purchase.
Note: messages.json is generated at runtime from hard coded defaults. The contents included in the
provided resource file does not represent the sum of the contents generated on startup.